+++ WowSrp/src/Header/HeaderData.cs
@@ -3,20 +3,23 @@ namespace WowSrp.Header
     /// <summary>
     ///     Container for header data.
     /// </summary>
-    public readonly struct HeaderData
+    public struct HeaderData
     {
         /// <summary>
         ///     Size of the message including the opcode field.
         /// </summary>
-        public uint Size { get; }
+        public uint Size { get; set; }
 
         /// <summary>
         ///     Opcode of the message.
         ///     Exact meaning depends on whether this is from the client or server.
         /// </summary>
-        public uint Opcode { get; }
+        public uint Opcode { get; set; }
 
-        internal HeaderData(uint size, uint opcode)
+        /// <summary>
+        ///     Constructor.
+        /// </summary>
+        public HeaderData(uint size, uint opcode)
         {
             Size = size;
             Opcode = opcode;